pyproject.toml
[project]
name = "ai-cr-service"
version = "0.1.0"
description = "AI Code Review Service for GitLab MR"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-gitlab>=5.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "redis>=5.2.0",
    "openai>=1.57.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_cr_service"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
